project/requirements.txt
Flask==2.0.1
pyodbc==4.0.30
matplotlib==3.4.2
azure-storage-blob==12.8.0
azure-identity==1.6.0
